refactor(p2p): Cleanup p2p code
